Conflicts:
	src/main/java/com/bonc/epm/paas/controller/ServiceController.java

+++ src/main/java/com/bonc/epm/paas/controller/ServiceController.java
@@ -9,6 +9,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -400,8 +401,7 @@ public class ServiceController {
 	 * @return
 	 */
 	@RequestMapping(value = { "service/add" }, method = RequestMethod.GET)
-	public String create(String imgID, String imageName, String imageVersion, String resourceName, Model model) {
-		
+	public String create(String imgID, String imageName, String imageVersion, String resourceName, Model model) {		
 		String isDepoly = "";
 		if (imageName != null) {
 			isDepoly = "deploy";
@@ -446,7 +446,12 @@ public class ServiceController {
                 for (int i = 0;i<countOfExposedPort;i++) {
                     PortConfig portConfig = new PortConfig();
                     portConfig.setContainerPort(String.valueOf(exposedPorts[i].getPort()));
-                    portConfig.setMapPort(String.valueOf(vailPortSet()));
+                    int randomPort = vailPortSet();
+                    if (-1 != randomPort) {
+                        portConfig.setMapPort(String.valueOf(randomPort));                       
+                    } else {
+                        new RuntimeException("获取端口失败！");
+                    }
                     tmpPortConfigs.add(portConfig);
                 }
                 return tmpPortConfigs;
@@ -458,8 +463,7 @@ public class ServiceController {
     @RequestMapping(value = { "service/getPortConfig.do" }, method = RequestMethod.GET)
     @ResponseBody
     public String getPortConfig(String imgID){
-        List<PortConfig> list = new ArrayList();
-        list=getBaseImageExposedPorts(imgID);
+        List<PortConfig> list=getBaseImageExposedPorts(imgID);
         Map<String,Object> map = new HashMap<String,Object>();
         map.put("data",list);
         return JSON.toJSONString(map); 
@@ -664,6 +668,9 @@ public class ServiceController {
 				portCon.setCreateDate(new Date());
 				portCon.setServiceId(service.getId());
 				portConfigDao.save(portCon);
+				
+				// 向map中添加生成的node端口
+				smalSet.add(Integer.valueOf(portCon.getContainerPort()));
 			}
 		}
 		//保存
@@ -761,45 +768,47 @@ public class ServiceController {
 		map.put("data", envVariables);
 		return JSON.toJSONString(map);
 	}
-	
 	/**
-	 * yuanpeng
-	 * 生成有效的PORTSET
-	 * @return int
-	 */
-	public int vailPortSet(){
-	    synchronized (this) {
-	        int stopFlag=0;
-	        int offset = kubernetesClientService.getK8sEndPort() - kubernetesClientService.getK8sStartPort();
-	        Set<Integer> bigSet = Stream.iterate(kubernetesClientService.getK8sStartPort(), item -> item+1)
-	                                    .limit(offset)
-	                                    .collect(Collectors.toSet());
-	        Set<Integer> tmpBigSet = bigSet;
-	        tmpBigSet.removeAll(smalSet);
-	        while (CollectionUtils.isEmpty(tmpBigSet)) {
-	            smalSet.clear();
-	            smalSet.addAll(portConfigDao.findPortSets());
-	            smalSet.remove(null);
-	            bigSet.removeAll(smalSet); 
-	            tmpBigSet = bigSet;
-	            stopFlag++;
-	            if (stopFlag > 100) {
-	                return -1;
-	                            }
-	                }
-	        Object[] obj =bigSet.toArray();
-	        int portSet=Integer.valueOf(obj[(int)(Math.random()*obj.length)]
-	                           .toString());
-	        smalSet.add(portSet);
-	        return portSet;
-        }
-	}
+    * 生成有效的PORTSET,回收端口
+    * @return int
+    */
+   public int vailPortSet(){
+       synchronized (this) {
+           int offset = kubernetesClientService.getK8sEndPort() - kubernetesClientService.getK8sStartPort();
+           Set<Integer> bigSet = Stream.iterate(kubernetesClientService.getK8sStartPort(), item -> item+1)
+                                       .limit(offset)
+                                       .collect(Collectors.toSet());
+/*           Set<Integer> tmpBigSet = bigSet;
+           tmpBigSet.removeAll(smalSet);
+           if (CollectionUtils.isEmpty(tmpBigSet)) {
+               smalSet.clear();
+               smalSet.addAll(portConfigDao.findPortSets());
+               smalSet.remove(null);
+               bigSet.removeAll(smalSet); 
+               return -1;
+           } else {
+               smalSet.addAll(portConfigDao.findPortSets());
+               smalSet.remove(null);
+               bigSet.removeAll(smalSet);
+           }*/
+           bigSet.removeAll(smalSet);
+           if (CollectionUtils.isEmpty(bigSet)) {
+               smalSet.addAll(portConfigDao.findPortSets());
+               smalSet.remove(null);
+               return -1;
+           }
+           Object[] obj =bigSet.toArray();
+           int portSet=Integer.valueOf(obj[(int)(Math.random()*obj.length)]
+                              .toString());
+           return portSet;
+       }
+   }
 
 	@RequestMapping(value = {"service/generatePortSet.do"} , method = RequestMethod.GET)
 	@ResponseBody
 	public String generatePortSet(){
 		Map<String, String> map = new HashMap<String, String>();
-		if(-1==vailPortSet()){
+		if(-1 == vailPortSet()){
 			map.put("ERROR","error");
 			return JSON.toJSONString(map);
 		}
@@ -814,12 +823,9 @@ public class ServiceController {
 	 * @return 
 	 */
 	@RequestMapping(value = { "service/removeSet.do" } , method = RequestMethod.GET)
-	@ResponseBody
-	public String removeSet(int set){
-	  Map<String, String> map = new HashMap<String, String>();
-		smalSet.remove(set);
-		map.put("status", "200");
-		return JSON.toJSONString(map);
+	public void removeSet(int set){
+	    log.info("移除的端口："+set);
+		/*smalSet.remove(set);*/
 	}
 	
 	/**
@@ -1087,7 +1093,16 @@ public class ServiceController {
 			map.put("status", "200");
 			serviceDao.delete(id);
 			envVariableDao.deleteByServiceId(id);
-			portConfigDao.deleteByServiceId(id);
+			
+			// 删除服务 释放绑定的端口
+			List<PortConfig> bindPort = portConfigDao.findByServiceId(id);
+			if (CollectionUtils.isNotEmpty(bindPort)) {
+			    for (PortConfig oneRow : bindPort) {
+			        smalSet.remove(Integer.valueOf(oneRow.getContainerPort()));
+			    }
+			    portConfigDao.deleteByServiceId(id);
+			}
+			
 			// 更新挂载卷的使用状态
 			if (!"0".equals(service.getVolName()) && StringUtils.isNotBlank(service.getVolName())) {
 			    this.updateStorageType(service.getVolName(), service.getServiceName());
